+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Improve datadog metrics by unifying span naming and adding more context.
+
 ## [v18.2.6](https://github.com/lexicalunit/spellbot/releases/tag/v18.2.6) - 2026-02-25
 
 ### Fixed